+++ src/backend/core/services/search.py
@@ -8,6 +8,7 @@ from core import enums
 
 from .embedding import embed_text
 from .opensearch import check_hybrid_search_enabled, opensearch_client
+from .reranking import rerank, should_rerank
 
 logger = logging.getLogger(__name__)
 
@@ -25,6 +26,7 @@ def search(  # noqa : PLR0913
     groups,
     tags,
     path=None,
+    rerank_requested=None,
 ):
     """Perform an OpenSearch search"""
     query = get_query(
@@ -37,7 +39,7 @@ def search(  # noqa : PLR0913
         tags=tags,
         path=path,
     )
-    return opensearch_client().search(  # pylint: disable=unexpected-keyword-arg
+    response = opensearch_client().search(  # pylint: disable=unexpected-keyword-arg
         index=",".join(search_indices),
         body={
             "_source": enums.SOURCE_FIELDS,  # limit the fields to return
@@ -58,6 +60,14 @@ def search(  # noqa : PLR0913
         # ignore_unavailable is not in the method declaration
         ignore_unavailable=True,
     )
+
+    if should_rerank(rerank_requested) and q != "*":
+        response["hits"]["hits"] = rerank(
+            query=q,
+            hits=response["hits"]["hits"],
+        )
+
+    return response
 
 
 # pylint: disable=too-many-arguments, too-many-positional-arguments