+++ src/backend/core/tests/test_search.py
@@ -5,6 +5,7 @@ Test suite for opensearch search service
 import logging
 import operator
 from json import dumps as json_dumps
+from unittest.mock import patch
 
 import pytest
 import responses
@@ -603,3 +604,58 @@ def test_search_filtering_by_query_path_and_tag():
 
     assert result["hits"]["total"]["value"] == len(documents_to_search)
     assert returned_ids == expected_ids
+
+
+# pylint: disable=too-many-arguments, too-many-positional-arguments
+@responses.activate
+@patch("core.services.search.rerank")
+@pytest.mark.parametrize(
+    "rerank_requested, reranker_enabled, rerank_called, caplog_messages",
+    [
+        (True, True, True, []),
+        (
+            True,
+            False,
+            False,
+            [
+                "Reranking was explicitly requested but the reranker is "
+                "disabled in settings. Reranking skipped."
+            ],
+        ),
+        (False, True, False, []),
+        (False, False, False, []),
+        (None, False, False, []),
+        (None, True, True, []),
+    ],
+)
+def test_search_rerank(  # noqa : PLR0913
+    mock_rerank,
+    rerank_requested,
+    reranker_enabled,
+    rerank_called,
+    caplog_messages,
+    settings,
+    caplog,
+):
+    """Test that rerank activation according to args and settings"""
+    settings.RERANKER_ENABLED = reranker_enabled
+
+    documents = bulk_create_documents(
+        [
+            {"title": "wolf", "content": "wolves live in packs and hunt together"},
+            {"title": "dog", "content": "dogs are loyal domestic animals"},
+            {"title": "cat", "content": "cats are curious and independent pets"},
+        ]
+    )
+    mock_rerank.return_value = documents
+    service = factories.ServiceFactory(name=SERVICE_NAME)
+    prepare_index(service.index_name, documents)
+
+    with caplog.at_level(logging.INFO):
+        search(
+            q="canine pet", rerank_requested=rerank_requested, **search_params(service)
+        )
+
+    assert mock_rerank.called == rerank_called
+    for message in caplog_messages:
+        assert message in caplog.text