+++ src/backend/core/views.py
@@ -346,6 +346,9 @@ class SearchDocumentView(ResourceServerMixin, views.APIView):
             List of public/authenticated documents the user has visited to limit
             the document returned to the ones the current user has seen.
             Built from linkreach list of a document in docs app.
+        rerank : bool, optional
+            Enable or disable reranking of results. If not specified, falls back to
+            the RERANKER_ENABLED setting.
 
         Returns:
         --------
@@ -356,7 +359,6 @@ class SearchDocumentView(ResourceServerMixin, views.APIView):
         # Get list of groups related to the user from SCIM provider (consider caching result)
         audience = self._get_service_provider_audience()
         user_sub = self.request.user.sub
-        groups = []
         params = schemas.SearchQueryParametersSchema(**request.data)
 
         # Get index list for search query
@@ -379,9 +381,10 @@ class SearchDocumentView(ResourceServerMixin, views.APIView):
             reach=params.reach,
             visited=params.visited,
             user_sub=user_sub,
-            groups=groups,
+            groups=[],
             tags=params.tags,
             path=params.path,
+            rerank_requested=params.rerank,
         )["hits"]["hits"]
         logger.info("found %d results", len(result))
         logger.debug("results %s", result)